Moving utils to dfm-data package
